+++ src/XrmMockupShared/Utility.cs
@@ -558,7 +558,6 @@ namespace DG.Tools.XrmMockup
                     criteria.Conditions.Any(c => EvaluateCondition(row, c));
         }
 
-
         private static bool EvaluateCondition(Entity row, ConditionExpression condition)
         {
             object attr = null;
@@ -689,7 +688,10 @@ namespace DG.Tools.XrmMockup
                     }
                     var x = int.Parse((string)values.First());
                     return now.Date <= date.Date && date.Date <= now.AddYears(x).Date;
-
+                case ConditionOperator.In:
+                    return values.Contains(attr);
+                case ConditionOperator.NotIn:
+                    return !values.Contains(attr);
                 default:
                     throw new NotImplementedException($"The ConditionOperator '{op}' has not been implemented yet.");
             }