+++ RELEASE_NOTES.md
@@ -1,3 +1,21 @@
+### 0.9.9-beta - 5 November  2018
+* Add support for 'In' and 'NotIn' in conditional expression for query expressions
+
+### 0.9.8-beta - September  2018
+* Fixed a bug where members of a team could not read any records owned by a team in the same Business Unit og child Business Unit
+
+### 0.9.7-beta - September 12 2018
+* Added target type Guid, fixed bug when reference did not have value.
+
+### 0.9.6-beta - August 16 2018
+* Fixed a bug in delete request where a related entity only had N:N relationship.
+
+### 0.9.5-beta - June 29 2018
+* Fixed a bug when accessing ParentPluginContext on a PluginContext.
+
+### 0.9.4-beta - June 22 2018
+* Fixed a bug where registered plugins are disabled when XrmMockup is initialized first time
+
 ### 0.9.3-beta - June 13 2018
 * Fixed a bug in ExecuteMultiple where IsFaulted was not set